cli: sv-epoch
