+++ rinex-cli/Cargo.toml
@@ -20,3 +20,4 @@ itertools = "0.10"
 plotly =  "0.8.3"
 colorous = "1"
 map_3d = "0.1.5"
+ndarray = "0.15"
