+++ rinex-cli/src/plot/context.rs
@@ -1,0 +1,15 @@
+use super::Plot; 
+
+pub struct Context {
+    nb_plots: usize,
+    plot: Plot, 
+}
+
+impl Context {
+    pub fn new() -> Self {
+        Self {
+            nb_plots: 0,
+            plot: Plot::new(),
+        }
+    }
+}
